split different gen config elements into files
Cuts up the massive `pkg/generate/config/config.go` file into separate
files for field, operation and resource. This patch contains no
functional changes; it is simply prep for work that will refactor the
FieldConfig and AdditionalFieldConfig structs.
